--- a/include/tmc/aw_resume_on.hpp
+++ b/include/tmc/aw_resume_on.hpp
@@ -35,6 +35,10 @@ public:
   /// When awaited, the outer coroutine will be resumed with the provided
   /// priority.
   inline aw_resume_on& with_priority(size_t Priority) {
+    // For this to work correctly, we must change the priority of the executor
+    // thread by posting the task to the executor with the new priority.
+    // Directly changing detail::.this_thread::this_task.prio is insufficient,
+    // as it doesn't update the task_stopper_bitsets.
     prio = Priority;
     return *this;
   }