--- a/include/tmc/detail/ex_cpu.ipp
+++ b/include/tmc/detail/ex_cpu.ipp
@@ -603,8 +603,8 @@ namespace detail {
 tmc::task<void> client_main_awaiter(
   tmc::task<int> ClientMainTask, std::atomic<int>* ExitCode_out
 ) {
-  ClientMainTask.resume_on(tmc::cpu_executor());
-  int exitCode = co_await std::move(ClientMainTask);
+  int exitCode =
+    co_await std::move(ClientMainTask.resume_on(tmc::cpu_executor()));
   ExitCode_out->store(exitCode);
   ExitCode_out->notify_all();
 }