### Breaking Changes
Require all co_await/detach/post expressions be on an rvalue. This emphasizes that tmc::task, and other awaitables that wrap tasks, are linear types - operations that transform them into another type (or void, by submitting it to the executor) "take ownership" by moving-from them.
```cpp
// Valid
co_await task();
co_await spawn(task());
co_await spawn(task()).run_early();
spawn(task()).detach();

// Not valid
auto x = task(); co_await x;
auto y = spawn(task()); co_await y;
auto z = spawn(task()).run_early(); co_await z;
auto w = spawn(task(); w.detach();

// Valid
auto x = task(); co_await std::move(x);
auto y = spawn(task()); co_await std::move(y);
auto z = spawn(task()).run_early(); co_await std::move(z);
auto w = spawn(task(); std::move(w).detach();
```

Return type of co_await is always an rvalue:
```cpp
// Valid
auto x = co_await task();
auto y = co_await spawn(task());
auto z = co_await spawn(task()).run_early();

// Not valid
auto& x = co_await task();
auto& y = co_await spawn(task());
auto& z = co_await spawn(task()).run_early();
```

### Performance Improvements
It was observed that rounding up the coroutine frame size in `operator new()` to the nearest 64 bytes leads to a small improvement in performance - because other threads can't false share with adjacent coroutine frames. For skynet. which had a starting coroutine frame size of 280 bytes, this meant rounding up to 320 bytes, and improved runtime on the EPYC 7742 from 730ms -> 695ms. Because this implemention also marks the operator new as `noexcept` (which means that the program will terminate if allocation fails), you must explicitly opt-in to this (roundup + noexcept) behavior by defining `TMC_CUSTOM_CORO_ALLOC`.

Then, as part of the implementation of the linear type rules above, all task/spawn awaitables were separated into a pre-await type and a post-await type (returned by operator co_await). This allows the compiler to reason about their separate lifetimes, and reduces the size of the coroutine frame. Additionally, for spawn_many(), the materialization of the supplied iterator into an array of tasks is delayed until await_suspend(). This causes the array of tasks to be created on the stack instead of in the coroutine frame. Overall, this reduces the raw coroutine frame size of skynet from 280 bytes to 184 bytes. This is then rounded-up by `operator new()` to 192 bytes (with `TMC_CUSTOM_CORO_ALLOC` enabled).

This also means that Clang is able to perform coroutine RVO!

TMC_FORCE_INLINE was applied to most await_suspend functions - I believe these should be inlined anyway as they result in only a few assembly instructions. It may be that these are not inlined automatically because Clang still decorates them with `noinline` as part of the bugfix for incorrect code generation (spilling coroutine variables to the coroutine frame after it was suspended). However, in Clang 18 that bug has been resolved so the `noinline` specifier should have been removed. This application of TMC_FORCE_INLINE reduces the runtime of `fib 42` from 440ms -> 410ms.

### Other
- Mark all builder functions as nodiscard.
- Preserve value category across builder function chains.
- Implement mixins to simplify reuse of builder functions / maintain value category / linear type rules.
- Implement into_unsafe_task to simplify internal handling of move-only tasks (rules for thee... and not for me!)